requirements.txt
# --- Web Framework & Server ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
requests>=2.31.0
aiofiles>=24.1.0

# --- The Swarm Brain (2026 Versions) ---
langchain>=0.3.0
langchain-core>=1.0.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langgraph>=1.0.7
pydantic>=2.11.2

# --- Specialized Tools ---
pypdf>=6.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1